requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
celery==5.2.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
colorlog==6.7.0
constantly==23.10.4
cryptography==42.0.5
Django==4.2.6
django-celery-beat==2.1.0
django-celery-results==2.4.0
django-compression-middleware==0.4.1
django-cors-headers==3.10.1
django-environ==0.9.0
django-extensions==3.2.3
django-filter==21.1
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
drf-nested-routers==0.93.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
gevent==23.9.0
greenlet==2.0.0
msgpack==1.0.8
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2023.4
redis==5.0.3
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==1.26.18
uvloop==0.19.0
uWSGI==2.0.22
whitenoise==6.6.0
zope.interface==6.2
zstandard==0.22.0
